Move Hotels.js into pages folder
