+++ src/pages/Hotels.js
